requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==1.0
Flask-paginate==0.1.5
Flask-Script==0.6.7
misaka==1.0.2
Pygments==1.6
Unidecode==0.04.13
beautifulsoup4==4.2.1
requests==1.2.3
pytz>=2013b
python-dateutil==2.1
